Frame/View Split and initial implementation of tree view selector
This has been a huge overhaul to both the frontend and backend architecture, and only now can I claim that this app has started to acquire some practical utility: essentially, each dataframe is associated with various views, with each view based on a particular lazyframe. The lazy evaluation of views mean that it would be very cheap to switch between views, and it would be easy to create new views based on existing ones. However, there's still some problems that need to be fixed for the tree view.